Convert to asbtract XYZ class
